+++ lib/widgets/dialogs/settings_dialog.dart
@@ -37,6 +37,7 @@ import '../../state/tabs_provider.dart';
 import '../../state/consent_provider.dart';
 import '../../state/info_safety_provider.dart';
 import '../../theme/app_theme.dart';
+import '../../theme/brand_logo.dart';
 import '../../utils/error_snackbar.dart';
 import '../../utils/log.dart';
 import '../../utils/url_launcher_util.dart';